backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Market data
yfinance==0.2.38
pandas==2.2.1
numpy==1.26.4

# News & sentiment
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3

# LLM providers
openai==1.20.0
anthropic==0.25.1
google-generativeai==0.5.0

# Caching (optional)
redis==5.0.3

# Scheduler
apscheduler==3.10.4

# HTTP client for testing
httpx==0.27.0
